Lots of animations tweaked to match the game. Added full synchro counter animation and logic to cards and card ui. Fixed some bugs. TODO: for some reason, attacks no longer deal damage to enemies.

+++ BattleNetwork/bnTextBox.cpp
@@ -156,7 +156,7 @@ void TextBox::ShowPreviousLine() {
 
 void TextBox::CompleteCurrentBlock()
 {
-  int newCharIndex = message.size() - 1;
+  size_t newCharIndex = message.size() - 1;
   int lastLine = lineIndex + GetNumberOfFittingLines();
 
   if (lastLine < this->lines.size()) {
